Move RawPoints to its own header, make it queryable to simplify other logic
